requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9      # required for FastAPI file uploads

# ── Document parsing ───────────────────────────────────────────────────────────
pymupdf>=1.23.0              # PDF parsing  (import fitz)
python-docx>=1.1.0           # DOCX parsing
lxml>=4.9.0                  # XML backbone used by python-docx

# ── Image handling ─────────────────────────────────────────────────────────────
Pillow>=9.0.0

# ── TTS ───────────────────────────────────────────────────────────────────────
gTTS>=2.5.0                  # primary TTS — Google TTS library (no API key)
pydub>=0.25.0                # MP3→WAV conversion used by the gTTS path
# System packages for audio conversion + offline fallback:
#   apt install ffmpeg espeak-ng

# ── ChartCaptioner model dependencies ─────────────────────────────────────────
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
nltk>=3.8.0
numpy>=1.24.0
tqdm>=4.65.0
